Refactor usePairLiquidityMiningCampaigns to it's own component - use getLowerTimeLimit and getTokenAmount
